backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
flask-swagger-ui==3.36.0
pykafka==2.8.0
pymongo==3.11.2
PyJWT==1.7.1
influxdb-client
influxdb==5.3.1
kafka-python==2.0.2
keras-tuner==1.0.3
Pebble==4.6.1
requests==2.26.0
scikit-learn==0.24.2
sklearn==0.0
tensorflow==2.5.0
